Now we're on commit cc9d44a; Changes from upstream we pulled:

- Minor: Username in channel points rewards redemption messages is now clickable. (#2673, #2953)
- Bugfix: Now deleting cache files that weren't modified in the past 14 days. (#2947)
- Bugfix: Middle mouse click no longer scrolls in not fully populated usercards and splits. (#2933)
- Bugfix: Fix bad behavior of the HTML color picker edit when user input is being entered. (#2942)

+++ src/messages/Image.cpp
@@ -406,9 +406,10 @@ void Image::actuallyLoad()
             buffer.open(QIODevice::ReadOnly);
             QImageReader reader(&buffer);
 
-            if (reader.size().width() * reader.size().height() *
-                    reader.imageCount() * 4 >
-                Image::maxBytesRam)
+            // use "double" to prevent int overflows
+            if (double(reader.size().width()) * double(reader.size().height()) *
+                    double(reader.imageCount()) * 4.0 >
+                double(Image::maxBytesRam))
             {
                 qCDebug(chatterinoImage) << "image too large in RAM";
